+++ engine/src/execution_kernels/LogicalProject.cpp
@@ -955,7 +955,7 @@ std::unique_ptr<ral::frame::BlazingTable> process_project(
 
         std::string name = named_expr.substr(0, named_expr.find("=["));
         std::string expression = named_expr.substr(named_expr.find("=[") + 2 , (named_expr.size() - named_expr.find("=[")) - 3);
-        expression = preprocess_expression_for_evaluation(expression, context, blazing_table_in->get_schema());
+        expression = preprocess_expression(expression, context, blazing_table_in->get_schema());
 
         expressions[i] = expression;
         out_column_names[i] = name;