+++ engine/tests/parser/expression_utils_test.cpp
@@ -281,7 +281,7 @@ TEST_F(ExpressionUtilsTest, concat_operator_with_empty_expressions)
 {
 	try {
 		std::string expression = "CONCAT()";
-		std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+		std::string out_expression = convert_nary_to_binary_concat(expression);
 
 		FAIL();
 	} catch(const std::exception& e) {
@@ -292,7 +292,7 @@ TEST_F(ExpressionUtilsTest, concat_operator_with_empty_expressions)
 TEST_F(ExpressionUtilsTest, concat_operator_wo_literals_expressions)
 {
 	std::string expression = "CONCAT($0, $1)";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 
 	EXPECT_EQ(out_expression, expression);
 }
@@ -300,14 +300,14 @@ TEST_F(ExpressionUtilsTest, concat_operator_wo_literals_expressions)
 TEST_F(ExpressionUtilsTest, concat_operator_with_one_literal_expressions)
 {
 	std::string expression = "CONCAT($0, '-ab25')";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 	EXPECT_EQ(out_expression, expression);
 }
 
 TEST_F(ExpressionUtilsTest, concat_operator_with_multiple_literal_expressions)
 {
 	std::string expression = "CONCAT(' - ', $1, ' : ')";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 	std::string expected_str = "CONCAT(CONCAT(' - ', $1), ' : ')";
 
 	EXPECT_EQ(out_expression, expected_str);
@@ -316,7 +316,7 @@ TEST_F(ExpressionUtilsTest, concat_operator_with_multiple_literal_expressions)
 TEST_F(ExpressionUtilsTest, concat_operator_using_cast_op)
 {
 	std::string expression = "CONCAT($0, ': ', CAST($1):VARCHAR, ' - ', $2)";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 	std::string expected_str = "CONCAT(CONCAT(CONCAT(CONCAT($0, ': '), CAST($1):VARCHAR), ' - '), $2)";
 
 	EXPECT_EQ(out_expression, expected_str);
@@ -325,16 +325,35 @@ TEST_F(ExpressionUtilsTest, concat_operator_using_cast_op)
 TEST_F(ExpressionUtilsTest, concat_operator_using_comma_as_literal)
 {
 	std::string expression = "CONCAT($0, ' , ', $2)";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 	std::string expected_str = "CONCAT(CONCAT($0, ' , '), $2)";
 
 	EXPECT_EQ(out_expression, expected_str);
 }
 
+TEST_F(ExpressionUtilsTest, concat_operator_using_binary_cast_op)
+{
+	std::string expression = "CONCAT($1, '-', CAST(+($0, 1)):VARCHAR)";
+	std::string out_expression = convert_nary_to_binary_concat(expression);
+	std::string expected_str = "CONCAT(CONCAT($1, '-'), CAST(+($0, 1)):VARCHAR)";
+
+	EXPECT_EQ(out_expression, expected_str);
+}
+
+TEST_F(ExpressionUtilsTest, concat_operator_into_another_operator) {
+	std::string expression =
+		"OPERATOR(CONCAT('Customer#000000', CAST($0):VARCHAR), 'Customer#0000001', "
+		"CONCAT('=order', CAST($901):STRING, 'another+op'))";
+	std::string out_expr = convert_nary_to_binary_concat(expression);
+	EXPECT_EQ(out_expr,
+		"OPERATOR(CONCAT('Customer#000000', CAST($0):VARCHAR), 'Customer#0000001', CONCAT(CONCAT('=order', "
+		"CAST($901):STRING), 'another+op'))");
+}
+
 TEST_F(ExpressionUtilsTest, concat_operator_inside_a_like_operator)
 {
 	std::string expression = "LIKE(CONCAT('Customer#000000', $0), 'Customer#0000001')";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 
 	EXPECT_EQ(out_expression, expression);
 }
@@ -342,7 +361,7 @@ TEST_F(ExpressionUtilsTest, concat_operator_inside_a_like_operator)
 TEST_F(ExpressionUtilsTest, multiple_concat_operator_inside_a_like_operator)
 {
 	std::string expression = "LIKE(CONCAT('abcd', '0000', $0), '001')";
-	std::string out_expression = convert_concat_expression_into_multiple_binary_concat_ops(expression);
+	std::string out_expression = convert_nary_to_binary_concat(expression);
 	std::string expected_str = "LIKE(CONCAT(CONCAT('abcd', '0000'), $0), '001')";
 
 	EXPECT_EQ(out_expression, expected_str);