+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - #1570 Fix build due to changes in rmm device buffer
 - #1576 Fix `CC`/`CXX` variables in CI
 - #1581 Fix latest cudf dependencies
+- #1582 Fix concat suite E2E test for nested calls
 - #1585 Fix for GCS credentials from filepath